use prepared SQL statements for named user sandbox objects

+++ src/test/resources/db/component/component_type_create_mysql.sql
@@ -6,7 +6,7 @@
 
 CREATE TABLE IF NOT EXISTS component_types
 (
-    component_type_id bigint NOT NULL COMMENT 'the internal unique primary index',
+    component_type_id smallint    NOT NULL COMMENT 'the internal unique primary index',
     type_name    varchar(255)     NOT NULL COMMENT 'the unique type name as shown to the user and used for the selection',
     code_id      varchar(255) DEFAULT NULL COMMENT 'this id text is unique for all code links,is used for system im- and export and is used to link coded functionality to a specific word e.g. to get the values of the system configuration',
     description  text         DEFAULT NULL COMMENT 'text to explain the type to the user as a tooltip; to be replaced by a language form entry'
