+++ src/test/resources/db/log/change_field_create_mysql.sql
@@ -6,7 +6,7 @@
 
 CREATE TABLE IF NOT EXISTS change_fields
 (
-    change_field_id   bigint           NOT NULL COMMENT 'the internal unique primary index',
+    change_field_id   smallint         NOT NULL COMMENT 'the internal unique primary index',
     table_id          bigint           NOT NULL COMMENT 'because every field must only be unique within a table',
     change_field_name varchar(255)     NOT NULL COMMENT 'the real name',
     code_id           varchar(255) DEFAULT NULL COMMENT 'to display the change with some linked information',
